Add leader config workaround for neovim
The current version of home-manager's neovim module inserts the plugin
config at the top of the init file, complicating the plugin
configuration process:
https://github.com/nix-community/home-manager/issues/4609
This commit adds an empty plugin at the top of the plugin list with the
sole purpose of setting the leader key early. This problem might get
fixed in a future release of home-manager.